renaming and restructuring
this commit contains better variable names and uses the Qt conventions. Widgets will be called widgets. It changes the implementation of multiple and single objects into the function that takes a list just calls the single function foreach element in the list.
Unused comments were removed. A tiny bugfix for left-mouse-button-handling uses the event instead of global App-variable.

The documentation was removed, since we don't need it in our project.